clean up instances part 2
