--- service-catalog-ui/src/components/ServiceInstanceDetails/ServiceInstanceInfo/ServiceInstanceInfo.scss
@@ -1,3 +1,0 @@
-.labels-list > * {
-  margin: 10px 10px 0 0 !important;
-}
