--- service-catalog-ui/src/helpers/instancesGQL/mutations.js
@@ -1,11 +1,0 @@
-import gql from 'graphql-tag';
-
-const deleteServiceInstance = gql`
-  mutation DeleteServiceInstance($name: String!, $namespace: String!) {
-    deleteServiceInstance(name: $name, namespace: $namespace) {
-      name
-    }
-  }
-`;
-
-export { deleteServiceInstance };
